dev_requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
-e git+https://github.com/agronholm/anyio.git@b5c21d040b9286921fdf5e0a6c723ce4ab5aa7e3#egg=anyio
aiohttp-cors==0.7.0
aiohttp==3.7.4.post0; python_version >= '3.6'
appdirs==1.4.4
appnope==0.1.2; sys_platform == 'darwin'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==20.3.0
backcall==0.2.0
black[d]==20.8b1
brotlipy==0.7.0
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
decorator==4.4.2
discord-py-slash-command==1.0.9.5
discord-py[voice]==1.6.0
discord.py-stubs==1.6.2
discord.py==1.6.0; python_full_version >= '3.5.3'
h11==0.12.0; python_version >= '3.6'
h2==3.2.0
hpack==3.0.0
httpcore==0.12.3; python_version >= '3.6'
httpx[http2]==0.17.1
hyperframe==5.2.0
idna==3.1
ipython-genutils==0.2.0
ipython==7.21.0
jedi==0.18.0; python_version >= '3.6'
loguru==0.5.3
multidict==5.1.0; python_version >= '3.6'
mypy-extensions==0.4.3
mypy==0.812
parso==0.8.1; python_version >= '3.6'
pathspec==0.8.1
peewee==3.14.4
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.18; python_full_version >= '3.6.1'
ptyprocess==0.7.0
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pygments==2.8.1; python_version >= '3.5'
pynacl==1.4.0
regex==2021.3.17
rfc3986[idna2008]==1.4.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0
tenacity==7.0.0
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
traitlets==5.0.5; python_version >= '3.7'
typed-ast==1.4.2
typing-extensions==3.7.4.3
uvloop==0.15.2
wcwidth==0.2.5
yarl==1.6.3; python_version >= '3.6'
